Using domains for the modules instead of making a module for every single table in the database, which is overkill. New modules; users, patients, physicians, appointments, prescriptions, symptom-condition-mapping
