+++ src/app.module.ts
@@ -9,16 +9,8 @@ import { UsersModule } from './users/users.module';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { PatientsModule } from './patients/patients.module';
 import { PhysiciansModule } from './physicians/physicians.module';
-import { SymptomsModule } from './symptoms/symptoms.module';
-import { ConditionsModule } from './conditions/conditions.module';
-import { PatientConditionsModule } from './patient-conditions/patient-conditions.module';
-import { PatientSymptomsModule } from './patient-symptoms/patient-symptoms.module';
 import { SymptomConditionMappingModule } from './symptom-condition-mapping/symptom-condition-mapping.module';
-import { PhysicianConditionsModule } from './physician-conditions/physician-conditions.module';
-import { PhysicianSymptomsModule } from './physician-symptoms/physician-symptoms.module';
-import { PatientPhysiciansModule } from './patient-physicians/patient-physicians.module';
 import { AppointmentsModule } from './appointments/appointments.module';
-import { MedicationModule } from './medication/medication.module';
 import { PrescriptionModule } from './prescription/prescription.module';
 
 @Module({
@@ -41,16 +33,8 @@ import { PrescriptionModule } from './prescription/prescription.module';
     UsersModule,
     PatientsModule,
     PhysiciansModule,
-    SymptomsModule,
-    ConditionsModule,
-    PatientConditionsModule,
-    PatientSymptomsModule,
     SymptomConditionMappingModule,
-    PhysicianConditionsModule,
-    PhysicianSymptomsModule,
-    PatientPhysiciansModule,
     AppointmentsModule,
-    MedicationModule,
     PrescriptionModule],
   controllers: [AppController],
   providers: [AppService, AiService],
