--- src/conditions/conditions.controller.ts
@@ -1,4 +1,0 @@
-import { Controller } from '@nestjs/common';
-
-@Controller('conditions')
-export class ConditionsController {}
