--- src/medication/medication.controller.ts
@@ -1,4 +1,0 @@
-import { Controller } from '@nestjs/common';
-
-@Controller('medication')
-export class MedicationController {}
