--- src/patient-conditions/patient-conditions.controller.ts
@@ -1,4 +1,0 @@
-import { Controller } from '@nestjs/common';
-
-@Controller('patient-conditions')
-export class PatientConditionsController {}
