--- src/patient-physicians/patient-physicians.controller.ts
@@ -1,4 +1,0 @@
-import { Controller } from '@nestjs/common';
-
-@Controller('patient-physicians')
-export class PatientPhysiciansController {}
