--- src/patient-symptoms/patient-symptoms.controller.ts
@@ -1,4 +1,0 @@
-import { Controller } from '@nestjs/common';
-
-@Controller('patient-symptoms')
-export class PatientSymptomsController {}
