--- src/patient-symptoms/patient-symptoms.service.spec.ts
@@ -1,83 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { PatientSymptomsService } from './patient-symptoms.service';
-import { getRepositoryToken } from '@nestjs/typeorm';
-import { PatientSymptoms } from './patient-symptom.entity';
-import { Repository } from 'typeorm';
-
-describe('PatientSymptomsService', () => {
-  let service: PatientSymptomsService;
-  let repository: Repository<PatientSymptoms>;
-
-  const mockRepository = {
-    find: jest.fn(),
-    findOneBy: jest.fn(),
-    delete: jest.fn(),
-  };
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [
-        PatientSymptomsService,
-        {
-          provide: getRepositoryToken(PatientSymptoms),
-          useValue: mockRepository,
-        },
-      ],
-    }).compile();
-
-    service = module.get<PatientSymptomsService>(PatientSymptomsService);
-    repository = module.get<Repository<PatientSymptoms>>(
-      getRepositoryToken(PatientSymptoms),
-    );
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-
-  describe('findAll', () => {
-    it('should return an array of patient symptoms', async () => {
-      const symptoms = [{ id: 1, patientId: 2, symptomId: 3 }];
-      mockRepository.find.mockResolvedValue(symptoms);
-
-      const result = await service.findAll();
-      expect(result).toEqual(symptoms);
-      expect(mockRepository.find).toHaveBeenCalled();
-    });
-  });
-
-  describe('findOne', () => {
-    it('should return a single patient symptom by ID', async () => {
-      const symptom = { id: 1, patientId: 2, symptomId: 3 };
-      mockRepository.findOneBy.mockResolvedValue(symptom);
-
-      const result = await service.findOne(1);
-      expect(result).toEqual(symptom);
-      expect(mockRepository.findOneBy).toHaveBeenCalledWith({ id: 1 });
-    });
-
-    it('should return null if no patient symptom is found', async () => {
-      mockRepository.findOneBy.mockResolvedValue(null);
-
-      const result = await service.findOne(999);
-      expect(result).toBeNull();
-      expect(mockRepository.findOneBy).toHaveBeenCalledWith({ id: 999 });
-    });
-  });
-
-  describe('remove', () => {
-    it('should delete a patient symptom by ID', async () => {
-      mockRepository.delete.mockResolvedValue({ affected: 1 });
-
-      await service.remove(1);
-      expect(mockRepository.delete).toHaveBeenCalledWith(1);
-    });
-
-    it('should handle non-existent patient symptom deletion gracefully', async () => {
-      mockRepository.delete.mockResolvedValue({ affected: 0 });
-
-      await service.remove(999);
-      expect(mockRepository.delete).toHaveBeenCalledWith(999);
-    });
-  });
-});
