--- src/physician-conditions/physician-conditions.controller.ts
@@ -1,4 +1,0 @@
-import { Controller } from '@nestjs/common';
-
-@Controller('physician-conditions')
-export class PhysicianConditionsController {}
