--- src/physician-symptoms/physician-symptoms.controller.ts
@@ -1,4 +1,0 @@
-import { Controller } from '@nestjs/common';
-
-@Controller('physician-symptoms')
-export class PhysicianSymptomsController {}
