+++ src/symptom-condition-mapping/condition.entity.ts
