+++ src/symptom-condition-mapping/symptom-condition-mapping.entity.ts
@@ -1,5 +1,5 @@
-import { Condition } from '../conditions/condition.entity';
-import { Symptom } from '../symptoms/symptom.entity';
+import { Condition } from './condition.entity';
+import { Symptom } from './symptom.entity';
 import { Entity, Column, PrimaryGeneratedColumn, OneToOne, JoinColumn} from 'typeorm';
 
 @Entity('symptom_condition_mapping')
