+++ src/symptom-condition-mapping/symptom-condition-mapping.module.ts
@@ -3,9 +3,15 @@ import { SymptomConditionMappingController } from './symptom-condition-mapping.c
 import { SymptomConditionMappingService } from './symptom-condition-mapping.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { SymptomConditionMapping } from './symptom-condition-mapping.entity';
+import { Symptom } from './symptom.entity';
+import { Condition } from './condition.entity';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([SymptomConditionMapping])],
+  imports: [
+    TypeOrmModule.forFeature([SymptomConditionMapping]),
+    TypeOrmModule.forFeature([Symptom]),
+    TypeOrmModule.forFeature([Condition])
+  ],
   controllers: [SymptomConditionMappingController],
   providers: [SymptomConditionMappingService]
 })
