+++ src/symptom-condition-mapping/symptom.entity.ts
