--- src/symptoms/symptoms.controller.ts
@@ -1,4 +1,0 @@
-import { Controller } from '@nestjs/common';
-
-@Controller('symptoms')
-export class SymptomsController {}
